requirements.txt
pandas
numpy
scikit-learn
selenium
pillow
urllib3
# -e .